+++ playwright-code-gen/src/domains/accounts/steps/accountStepDefinitions.ts
@@ -1,10 +1,10 @@
 import { Given, When, Then } from '@cucumber/cucumber';
-import { TestData } from '../../../shared/utilities/testData';
-import { getAccountsPage, getAccountSteps } from '../../../shared/utilities/hooks';
+import { TestData } from '../../../shared/utilities/test-data';
+import { getAccountsPage, getAccountUsecase } from '../../../shared/utilities/hooks';
 
 Given('I am on the accounts page', async function() {
-  const accountSteps = getAccountSteps();
-  await accountSteps.navigateToAccountsPage();
+  const accountsPage = getAccountsPage();
+  await accountsPage.navigate();
 });
 
 When('I click on the {string} button', async function(buttonText: string) {
@@ -17,18 +17,18 @@ When('I click on the {string} button', async function(buttonText: string) {
 });
 
 When('I fill in the account form with the following details:', async function(dataTable) {
-  const accountsPage = getAccountsPage();
+  const accountUsecase = getAccountUsecase();
   const testAccount = TestData.getTestAccount();
-  await accountsPage.fillAccountForm(testAccount);
+  await accountUsecase.createAccount(testAccount);
 });
 
 Then('I should see the account {string} in the accounts list', async function(accountName: string) {
-  const accountSteps = getAccountSteps();
-  await accountSteps.verifyAccountCreated(accountName);
+  const accountUsecase = getAccountUsecase();
+  await accountUsecase.verifyAccountExists(accountName);
 });
 
 Then(/^the total balance should include \$?([\d,]+\.\d{2})$/, async function (amountStr: string) {
     const cleaned = parseFloat(amountStr.replace(/,/g, ''));
-    const accountSteps = getAccountSteps();
-    await accountSteps.verifyTotalBalance(cleaned);
+    const accountUsecase = getAccountUsecase();
+    await accountUsecase.verifyTotalBalance(cleaned);
   });
