+++ playwright-code-gen/src/shared/utilities/world.ts
@@ -1,6 +1,6 @@
 import { BaseWorld } from './base-world';
 import { AccountsPage } from '../../domains/accounts/pages/AccountsPage';
-import { AccountSteps } from '../../domains/accounts/steps/accountSteps';
+import { AccountUsecase } from '../../domains/accounts/usecases/AccountUsecase';
 
 /**
  * World class to encapsulate test context and state
@@ -10,7 +10,7 @@ import { AccountSteps } from '../../domains/accounts/steps/accountSteps';
 export class World extends BaseWorld {
   // Domain-specific page objects and steps
   public accountsPage!: AccountsPage;
-  public accountSteps!: AccountSteps;
+  public accountUsecase!: AccountUsecase;
 
   constructor() {
     super();
@@ -27,6 +27,6 @@ export class World extends BaseWorld {
     
     // Initialize domain-specific page objects and steps
     this.accountsPage = new AccountsPage(this.getPage());
-    this.accountSteps = new AccountSteps(this.accountsPage);
+    this.accountUsecase = new AccountUsecase(this.accountsPage);
   }
 }
