+++ playwright-code-gen/src/tests/accounts.spec.ts
@@ -1,29 +1,29 @@
 import { test, expect } from '@playwright/test';
-import { TestData } from '../shared/utilities/testData';
 import { AccountsPage } from '../domains/accounts/pages/AccountsPage';
-import { AccountSteps } from '../domains/accounts/steps/accountSteps';
+import { AccountUsecase } from '../domains/accounts/usecases/AccountUsecase';
+import { TestData } from '@/shared/utilities/test-data';
 
 test.describe('Account Management', () => {
   let accountsPage: AccountsPage;
-  let accountSteps: AccountSteps;
+  let accountUsecase: AccountUsecase;
 
   test.beforeEach(async ({ page }) => {
     accountsPage = new AccountsPage(page);
-    accountSteps = new AccountSteps(accountsPage);
+    accountUsecase = new AccountUsecase(accountsPage);
   });
 
   test('should create a new account successfully', async () => {
     // Given I am on the accounts page
-    await accountSteps.navigateToAccountsPage();
+    await accountsPage.navigate();
 
     // When I create a new account
     const testAccount = TestData.getTestAccount();
-    await accountSteps.createNewAccount(testAccount);
+    await accountUsecase.createAccount(testAccount);
 
     // Then I should see the account in the list
-    await accountSteps.verifyAccountCreated(testAccount.name);
+    await accountUsecase.verifyAccountExists(testAccount.name);
     
     // And the total balance should be updated
-    await accountSteps.verifyTotalBalance(testAccount.balance);
+    await accountUsecase.verifyTotalBalance(testAccount.balance);
   });
 });
